+++ packages/oauth2/src/authorization-request/fetch-authorization-response.ts
@@ -1,12 +1,13 @@
 import { CallbackContext } from "@openid4vc/oauth2";
 import { createFetcher } from "@openid4vc/utils";
 import {
+  CONTENT_TYPES,
+  HEADERS,
   UnexpectedStatusCodeError,
   ValidationError,
   hasStatusOrThrow,
 } from "@pagopa/io-wallet-utils";
 
-import { CONTENT_TYPES, HEADERS } from "../constants";
 import { PushedAuthorizationRequestError } from "../errors";
 import {
   PushedAuthorizationRequestSigned,
@@ -69,7 +70,7 @@ export async function fetchPushedAuthorizationResponse(
       options.pushedAuthorizationRequestEndpoint,
       {
         body: new URLSearchParams({
-          clientId: options.pushedAuthorizationRequestSigned.client_id,
+          client_id: options.pushedAuthorizationRequestSigned.client_id,
           request: options.pushedAuthorizationRequestSigned.request,
         }),
         headers: {