+++ packages/oauth2/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @pagopa/io-wallet-oauth2
 
+## 0.6.0
+
+### Minor Changes
+
+- b2a7475: Add credential request implementation
+
+### Patch Changes
+
+- Updated dependencies [b2a7475]
+  - @pagopa/io-wallet-utils@0.6.0
+
+## 0.5.1
+
+### Patch Changes
+
+- c9e7651: Fix: client_id query param in PAR request
+  - @pagopa/io-wallet-utils@0.5.1
+
 ## 0.5.0
 
 ### Minor Changes