+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pagopa/io-wallet-utils
 
+## 0.6.0
+
+### Minor Changes
+
+- b2a7475: Add credential request implementation
+
+## 0.5.1
+
 ## 0.5.0
 
 ## 0.4.2